feat: 🎸 change format of /rows to fix the features order
Note that the datasets library does not
officially support a way to specify the columns order, but relying on
info.features gives the dataset maintainer the ability to change this
order if they want.

Effectively, Dict preserves the order of the items, while JSON does not:
we have to pass the features in an Array to ensure the order is
preserved.

BREAKING CHANGE: 🧨 format change