No cache expiration (#59)
* refactor: 💡 remove dead code

* feat: 🎸 Don't expire cache entries

Cache entries now live forever, until they are deleted or refreshed. The
max-age header sent to the client is now a hint, without relation with
the freshness of the cached response. By the way, the technical
endpoints (/valid, /cache, /cache-reports) are no more memoized.

* refactor: 💡 remove expire feature in memoize

As we don't use it anymore, we simplify the code by removing the ability
to set an expire time in "memoize()" function. Also: remove other
unneeded arguments of that function.