Double limit on size (#162)
* feat: 🎸 truncate the rows by size before putting in cache

Remarks: 1. this duplicates the code and processing (in cache.io, the
same is done when preparing the answer. It's ok temporarily, until the
cache is clean, then we might want to remove from cache.io). 2. ideally,
this loop could be done before to avoid downloading and processing rows
that will be removed later, in row.py. But I'm not sure it's possible or
a good idea, since in streaming mode it would mean doing a network
request per row, or add complexity to the code to use small size
batches.

* fix: 🐛 fix type