Simplify cache by dropping two collections (#202)
* docs: ✏️ add backup/restore to migration instructions

* feat: 🎸 pass the max number of rows to the worker

* feat: 🎸 delete the 'rows' and 'columns' collections

instead of keeping a large collection of rows and columns, then compute
the response on every endpoint call, possibly truncating the response,
we now pre-compute the response and store it in the cache. We lose the
ability to get the original data, but we don't need it. It fixes #197.
See
https://github.com/huggingface/datasets-preview-backend/issues/197#issuecomment-1092700076.

BREAKING CHANGE: 🧨 the cache database structure has been modified. Run
20220408_cache_remove_dbrow_dbcolumn.py to migrate the database.

* style: 💄 fix types and style

* docs: ✏️ add parameter to avoid error in mongodump

* docs: ✏️ mark ROWS_MAX_BYTES and ROWS_MIN_NUMBER as worker vars