Add datasets-server-worker to the Kube cluster (#236)
* docs: ✏️ add doc about AWS configure, eks, ecr

* docs: ✏️ add doc about kubernetes

* docs: ✏️ reorganize the files

* docs: ✏️ add tools

* feat: 🎸 add the structure for the datasets-server helm chart

* feat: 🎸 initial helm chart

* fix: 🐛 fix helm chart thanks to @XciD comments

* feat: 🎸 add workers

* feat: 🎸 fix NFS server

* test: 💍 disable a test for now

* fix: 🐛 remove file!

* feat: 🎸 fix images, and remove nfs for now

* feat: 🎸 never stop the worker

* feat: 🎸 upgrade the worker image (loop)

also: fix the dev configuration (typo on the name)

* fix: 🐛 name the objects with the chart + release

this way, two instances of datasets-server can run on the same
namespace, with different domain names. Fixes
https://github.com/huggingface/datasets-server/pull/227#issuecomment-1120715631

* Nfs (#242)

* docs: ✏️ clarify doc about Release and Instance name

* fix: 🐛 use the release name to name all the objects

this way, mongo, which prefixes its name with the release, appears in
the same "group" of pods when ordered alphabetically

* fix: 🐛 remove "public," from cache-control

it is not needed (public is only meant to force public caching for
authenticated requests, which is not the case). And it prevented nginx
to correctly cache the responses based on cache-control.

* refactor: 💡 create one template directory per service

Also: use a static name for the containers (see
https://github.com/huggingface/datasets-server/commit/7abd07dc0046410adbbceda6684ea14c5d0f7b48#r73243800)

* feat: 🎸 ignore cpu and ram tests if MAX set to 0

Also: add doc to allow using HF_DATASETS_CACHE and HF_MODULES_CACHE to
set the datasets library cache directories.

* feat: 🎸 add NFS for /assets and dataset library /cache

Also: remove the tests for the load (cpu/ram) in the workers. Note: we
run the containers with user 1000 and group 3000, and use an
initContainer to ensure the mounted volumes from NFS use these user and
group. See https://github.com/huggingface/datasets-server/issues/241 for
a possible security improvement on this.

* fix: 🐛 avoid alpine image, and fix the version

see https://www.google.com/search?q=why+not+use+alpine for why not using
alpine

* fix: 🐛 use a sanitized version of .Release.Name

because the length of the k8s names is limited

* refactor: 💡 move the mongodb URL to a dedicated variable

* fix: 🐛 use Recreate strategy for workers

* docs: ✏️ add a one-liner to create the secret