Store and compare worker+dataset repo versions (#624)
* feat: 🎸 add a version to the workers and compare utils

* style: 💄 fix style

* feat: 🎸 increase version

* feat: 🎸 upgrade libqueue and upgrade version to 1.0.0

also: test the version

* feat: 🎸 add skipped state and should_skip_job to the worker

increase version

* feat: 🎸 upgrade to libqueue 0.4.4

* feat: 🎸 add Worker.compare_major_version

and remove Worker.is_major_version_lower_than_worker

* feat: 🎸 add two fields to the cache collections

worker_version and dataset_git_revision

* feat: 🎸 process a job only if needed

A job is skipped if a cache entry already exists, if it is successful,
if the worker version had the same major number and if the dataset git
revision was the same.

* feat: 🎸 process a job only if needed

A job is skipped if a cache entry already exists, if it is successful,
if the worker version had the same major number and if the dataset git
revision was the same.

* feat: 🎸 upgrade libqueue and libcache

it will change the format of the cache reports, by adding the
worker_version and dataset_git_revision fields

* feat: 🎸 fix type

* feat: 🎸 update libcache

* feat: 🎸 allow multiple waiting jobs with the same arguments

it makes the logic simpler. The only restriction is that we can have
only one started job with the same set of arguments (type, dataset,
config, split). It's the job of the worker to check if the job should be
computed or not (worker.should_skip_job)

* feat: 🎸 update libqueue

* feat: 🎸 update libqueue and libcache

* feat: 🎸 upgrade docker images