feat: 🎸 add concept of Resource (#784)
* feat: 🎸 add concept of Resource

* ci: 🎡 fix unit tests for the job

* refactor: 💡 remove dead code

* test: 💍 give more time to avoid timeout to mongodb

* refactor: 💡 rename resource into resources

(I had an error from inside ffspec, seems like there is a name conflict
- weird!)

* refactor: 💡 use InitVar to avoid mutating fields

This way we can alsp ensure that the storage directory is not None

* fix: 🐛 release the resources on shutdown

they were released just after the app has been returned. But we want the
resources to be released when the app is shut down instead.

Note that as we need the resources (mainly the log, and the assets
directory for the prometheus endpoint), we need to allocate them at the
start of the function, instead of using the "on_startup" parameter of
the Starlette constructor.

* fix: 🐛 avoid using InitVar since they are hard to use

they are only available in __init__ and __post_init__. We don't want
that in the Resource inherited objects

* test: 💍 fix parameter name

* fix: 🐛 no need to call .allocate, it's done in __post_init__

* feat: 🎸 use primitive parameters, add release, add tests

* style: 💄 fix style

* refactor: 💡 rename AssetsDirectory to AssetsStorageAccess

to make it clear that the resource is the access to the storage, not the
assets directory in itself (which must be persisted and will not be
deleted when the ressource is released). Thanks @lhoestq for the idea.

* fix: 🐛 allocate assets storage for all the workers

It fixes the issue raised by @lhoestq at
https://github.com/huggingface/datasets-server/pull/784#discussion_r1100498336.

Note that it's a bit radical: all the workers now require write access
to the assets storage directory, even if they will not use it.
I think it's OK since I will be working very soon on
https://github.com/huggingface/datasets-server/issues/737, which will
merge all the workers in only one worker.

* refactor: 💡 change env var names for coherence

fixes
https://github.com/huggingface/datasets-server/pull/784/files/2301eb9e528afd4e46414f05f65e032831de6ea5#r1101583351

* fix: 🐛 missing env var

* ci: 🎡 fix env var

* refactor: 💡 replace LogResource with init_logging

As we don't allocate/release anythin, it does not fit inside the concept
of Resource

* feat: 🎸 add init_assets_directory method

* refactor: 💡 remove StorageResource + remove mongo.py

Storage does not need to be a resource. Also: remove the unnecessary
MongoConnection class

* fix: 🐛 fix details to make the CI pass

* test: 💍 fix tests

* refactor: 💡 remove dead code

* test: 💍 remove dead code

* style: 💄 fix style