Add /sse/hub-cache to update the Hub's backend cache (#1784)
* feat: 🎸 upgrade mongoengine and pymongo

it will be needed to use motor for sse-api (to be added to the code).
From what I checked in the changelog, nothing should be breaking

* feat: 🎸 add services/sse-api

empty, for now. It will provide the /sse/hub-cache endpoint

* feat: 🎸 add /metrics, setup github actions, docker, helm

* fix: 🐛 fix the docker compose files

* fix: 🐛 fix env var

* fix: 🐛 add missing env vars

* docs: fix doc string

* fix: 🐛 8084 seems to be already used in CI. Trying 8085

* feat: 🎸 add a toy SSE endpoint

* feat: 🎸 use a pub/sub system (still a toy example with numbers)

* feat: 🎸 set replicaset on the local mongodb container

it's needed to "watch" the changes; required for the sse service

* feat: 🎸 watch the cache collection updates

* style: 💄 fix mypy details

* style: 💄 remove unused import

* feat: 🎸 install motor

* ci: 🎡 try to setup replicaset on mongo in CI

* ci: 🎡 try to fix mongo launch

* ci: 🎡 install mongosh

* feat: 🎸 create SSE for existing cache entries, before updates

Note that the events created by the two loops might be mixed. I think
it's not really an issue, we can just ignore "init" operations received
after other operations, for the same dataset.

* feat: 🎸 setup replicaset in docker compose file

* fix: 🐛 fix the mongo URL, and wait for mongo to be healthy

* docs: ✏️ add doc on /sse/hub-cache

and remove unused variables

* ci: 🎡 remove unused variables

* ci: 🎡 fix ci: no need for manual mongosh and replicaset install